[beauty] include results_filename_affix in name of parameters.m file
Each time we run a simulation, the `parameter` file gets added to the output. The current naming is time-dependent to avoid overwrites, but in the case of running multiple simulations with varying parameters (e.g., intensities) in the same folder, it is not too helpful, e.g., `sub-001_parameters_11_06_2024_163206_990.mat`. Now the suffix (which usually specifies the setup) is included and timing info is sparsened.