[debug] only save medium properties as nifty for skull and layered sims
Medium properties are most informative for varied tissue types. They do not change for an individual, such that it may not require us to append a suffix that specifies targets , stimulation parameters etc.. However, they could previously get overwritten if a water simulation was run posthoc. This should ensure this does not happen.